+++ src/menu/Image.h
@@ -7,8 +7,8 @@ namespace Symp {
 
 class Image : public GuiComponent {
 public:
-	Image( const char* filePath, int iPosX, int iPosY);
-	Image( const char* filePath, int iPosX, int iPosY, int iScale);
+	Image( const char* filePath, float iPosX, float iPosY);
+	Image( const char* filePath, float iPosX, float iPosY, float iScale);
 	~Image();
 	virtual void update();
